refactor api wallets endpoints to own file
