+++ src/cryptoadvance/specter/templates/includes/tx-data.html
@@ -44,7 +44,7 @@
         }
         
         async fetchRawTx(txid) {
-            let url = `{{ url_for('wallets_endpoint.decoderawtx', wallet_alias='WALLET_ALIAS') }}`.replace("WALLET_ALIAS", this.wallet);
+            let url = `{{ url_for('wallets_endpoint_api.decoderawtx', wallet_alias='WALLET_ALIAS') }}`.replace("WALLET_ALIAS", this.wallet);
             var formData = new FormData();
             formData.append('txid', txid);
             formData.append('csrf_token', '{{ csrf_token() }}');
@@ -257,7 +257,7 @@
         }
         
         async fetchAddressIsMine(address) {
-            let url = `{{ url_for('wallets_endpoint.addressinfo', wallet_alias='WALLET_ALIAS') }}`.replace("WALLET_ALIAS", this.wallet);
+            let url = `{{ url_for('wallets_endpoint_api.addressinfo', wallet_alias='WALLET_ALIAS') }}`.replace("WALLET_ALIAS", this.wallet);
             var formData = new FormData();
             formData.append('address', address)
             formData.append('csrf_token', '{{ csrf_token() }}');
